stripping weaken property from bolas
bolas cause slowdown, no need for them to cause knockdown/stun too. pair them with a weapon to deal with a target, on their own they are simply a tool to slow someone down. this should be fine since legion can move very quickly anyhow, so should not be reliant on a stun so much as snaring their enemy and then getting hits in.