Update synths.dm
stripped out military synth variant.
synths are basically just humans which can repair with welding/wires but also need synthflesh to repair damaged externals. idea isn't for them to play as a hostile force but more of a 'distant' human, they have some advantages like no hunger, radiation immunity, toxin immunity etc and can't be dismembered but also have downsides - EMPs will mess them up bad, their skin falls off and people might not be friendly to them, and they have a naturally frail constitution due to the majority of their body being designed so that a disguise can be wrapped over the top of it, not a rigid structure designed for life.